Moved GraphvizVersion, removed DotFailure
