+++ src/main/java/org/schemaspy/output/diagram/graphviz/GraphvizVersion.java
@@ -18,7 +18,7 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
-package org.schemaspy.util;
+package org.schemaspy.output.diagram.graphviz;
 
 import java.util.ArrayList;
 import java.util.List;
